requirements.txt
chromadb==0.3.21
langchain==0.0.154
openai==0.27.2
openapi-schema-pydantic==1.2.4
prompt-toolkit==3.0.38
pypdf==3.8.1
streamlit==1.20.0
tiktoken==0.3.3
tokenizers==0.13.3
python-dotenv==1.0.0
pdfminer.six==20221105